% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibass_app.R
\name{available_apps}
\alias{available_apps}
\title{List available apps in \{vibass\} package.}
\usage{
available_apps()
}
\value{
Character vector.
}
\description{
App codes that are available for use in [vibass_app()].
}
\examples{
available_apps()
}
