% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibass_app.R
\name{available_apps_message}
\alias{available_apps_message}
\title{Return a string with the apps available.}
\usage{
available_apps_message()
}
\value{
Character string.
}
\description{
Return a string with the apps available.
}
\keyword{internal}
