% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_table.R
\name{summary_table}
\alias{summary_table}
\title{Print a standardised summary table}
\usage{
summary_table(mean, var, quant, ic95, prop0 = NULL, prop1 = NULL, label)
}
\arguments{
\item{mean}{Real.}

\item{var}{Real.}

\item{quant}{Numeric vector.}

\item{ic95}{Numeric vector.}

\item{prop0}{Real.}

\item{prop1}{Real.}

\item{label}{Character. Name of the summarised variable .}
}
\value{

}
\description{
Make a table of several summary statistics with proper formatting.
}
\details{
The table includes the mean, variance and standard deviation, a vector of 3
quantiles at 0.05, 0.50 and 0.95, a 95% centred confidence interval and a
numeric value to be interpreted as a proportion above 0 and 1. All values are
rounded to 2 significant digits.
}
\examples{
summary_table(mean = 1, var = 1, quant = 1:3, ic95 = 4:5, prop1 = .6, label = "test")
}
