% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\docType{data}
\name{Weights}
\alias{Weights}
\title{Weights of children}
\format{
An object of class \code{"data.frame"}.
}
\source{
To be added.
}
\usage{
data("Weights")
}
\description{
Data for the VIBASS session on linear models. The dataset include
data bout children, that includes weight, height and some other variables.
}
\examples{
data(Weights)
summary(Weights)

# ML estimates
lmW <- lm(weight ~ age, data = Weights)
summary(lmW) 
}
\references{
To be added.
}
\keyword{datasets}
