% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibass_app.R
\name{vibass_app}
\alias{vibass_app}
\title{VIBASS interactive apps.}
\usage{
vibass_app(x = NULL)
}
\arguments{
\item{x}{integer or character interpretable as integer. See
[available_apps()] for valid options.}
}
\value{
With no arguments, displays a list of available app codes.
   Otherwise, lauches the corresponding Shiny app.
}
\description{
Launches the Shiny interactive applications for the practicals.
}
\examples{
vibass_app()
}
