% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vibass_app.R
\name{vibass_app}
\alias{vibass_app}
\title{VIBASS interactive apps.}
\usage{
vibass_app(x = 1, bg = FALSE)
}
\arguments{
\item{x}{integer or character interpretable as integer. 1-2. Practical id.}
}
\description{
Launches the Shiny interactive applications for the practicals.
}
