% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Weights.R
\docType{data}
\name{Weights}
\alias{Weights}
\title{Weights of children}
\format{
An object of class \code{"data.frame"}.
}
\source{
VIBASS Team.
}
\usage{
data("Weights")
}
\description{
Data for the VIBASS session on linear models. This is a simulated dataset
that includes data about children. The variables in the dataset are:
}
\details{
\itemize{
 \item age. Age (in years).
 \item vegetables. Measure of vegetables consumption.
 \item weight. Weight (in kg).
 \item sex. Girl or Boy.
 \item height. Height (in cm).
 \item race. Assian, Black or European.
}
}
\examples{
data(Weights)
summary(Weights)

# ML estimates
lmW <- lm(weight ~ age, data = Weights)
summary(lmW) 
}
\keyword{datasets}
